Introduce new 'show' option for control labels
Introduces a new configuration option in ui schemas for labels of
controls. In particular a 'show' option is added which allows to hide labels.

Instead of simply a string, a label for a control can now be defined as string,
boolean or {text:string, show:boolean} object. If not specified, 'true' for
show and the 'beautified schemapath' for text are assumed.

Includes testcases.