TreeRenderer and Drag & Drop Rework - unify anyOf Lists
Changed the tree to render multiple containment properties that belong
to one property in the data into the same list. This is the case when a
property allows multiple different types by using `anyOf` in the JSON
schema.

Also adapted Drag and Drop to deal with this. When a tree element is
dropped at an illegal target list, it is returned to its origin.
Consequently, it is made sure to only move tree elements to valid
locations.

Also adapted the test cases of the tree renderer and the drag and drop
handlers to the new behavior and added some new test cases.