Extract Fields from base controls (#716)
Created field renderer and a control which uses a delegate to fields

This commit defines the basic fields and reuses the tester we already
had for controls.
Also a control is defined which uses a delegate component to retrieve
the correct field.

This change also implies the following, additional changes:

* Removed base controls and adapted materialize controls to moved tester

As we now have a common renderer, which is able to dispatch to inner
fields, we can remove the individual base controls.
Also updated the materialize renderer to use the testers from fields
instead from the controls

* Fixed common control to be fully hidden and not to only hide the input

* Added explicit example for default renderer. Also fixed rating.

This adds a new example folder which is configured to provide only the
default renderers.
This also fixes the rating control in the examples by giving each entry
a key.

* Fixed failing test, due to wrong type

* Fixed naming; removed hidden attribute and state from fields

This fixes the naming and removes the unneeded state in the field
components. Thus the field components are now converted to functional
components.
Also removed the hidden attribute from the fields.