Add JsonEditor as package
Added a new package for the JsonEditor.
Bundle own SchemaService in JsonEditor so that the editor won't be
affected when it is removed from Core.
Adapt non-react tree master detail renderer to be its own element: it is
not rendered as a renderer by JsonForms anymore. Instead it is
configured directly by the editor element.
Besides the ui schema and styling registry, the JsonEditor is now
independent of the global JsonForms object.

The Ecore demo runs in general but still has some issues to fix:
* The tree dialog and the toolbar are not styled with Materialize. The
problem is: If materialize.css is loaded by the example editor, the
renderers in the detail are displayed incorrectly: Probably because they
use MaterialUI and some styles are then applied two times.

TODO:
* fix material CSS
* implement reference/attribute renderer as react renderers
* cleanup unneeded code

Json Editor - Styling Fix

Fixed the materialize CSS of the tree and the toolbar of the example
editor by copying the needed parts of materialize.css (mainly buttons
and nav).

Note: Cannot use the whole materialize.css because this leads to display
errors in combination with the JsonForms material renderers that do not
used materialize.css but Material UI.
When the editor and the tree are ported to React and Material UI this is
no longer needed.

Tried to use old tree renderer with detail react component

Should be obsolete after using React based tree.

tree master detail runs as react component in jsoneditor

more work on tree in jsoneditor