Fixed D&D cancelled prematurely in Chrome
There is a bug in Chrome that might cancel a HTML5 drag and drop
if the DOM is changed during the "begin drag" event.
Therefore, The CSS highlighting whether a list in the tree is a
valid or invalid drop target is set with a delay after starting
the drag.