Fix hasRef performance issues.
When running with the `@jsonforms/react` there is a significant performance issue when rendering a form with a lot of fields. I believe I have tracked this down to [excessive calls to `hasRefs`](https://github.com/eclipsesource/jsonforms/blob/master/packages/react/src/JsonForms.tsx#L78). As I understand the code, that deferred call to resolve the refs on the schema, only needs to happen once on the root `JsonForms` component, as it will pass/provide the already resolve schema down to children, however, all of the children are also performing the `hasRefs` check, even though there will be no refs as the root `JsonForms` component has resolve them. `hasRefs` is an expensive call, especially when run on a large form. I was able to drop render time from 30s to 3 seconds by removing that call.

Fixes #1208