Use schema title attribute where appropriate
The label utility will now return the schema title if no other label was
defined via the ui schema.

Within the ui schema a label can be optionally defined for each control.
Previously if there was none, we fell back to auto-generating a label
from the given control scope. Now we will check the schema title
beforehand and return that if no label was defined for the control.

This also affects the ui schema generation which will no longer
auto-generate labels based on each control's scope.

Includes testcases.