Fix moving table array fields
The width of table columns is determined by its content. By using
'Hidden' for adornments the width of the contained butiton was
not considered initially, but only when shown on hover. This lead to a
width increase for the control and therefore the whole column. We now use
'visibility: hidden' instead as this doesn't change the layout.