Don't cut off labels in MaterialInputControl
Adjust the custom classes for MaterialInputControl's InputLabel to not
cut off characters.

We use a special style to increase the InputLabel's length when
shrinked as well as to avoid line breaks by rendering ellipses instead.
With an update of material ui the custom styles collided with the
default ones and lead to characters like 'ypq' etc. to be
cut off. This cutoff is now avoided. Also the width of the InputLabel is
adjusted to respect the input adornment when not shrinked.