Add enum autocomplete renderers
Adds the `MuiAutocomplete' component which is used by the new
autocomplete enum and oneOfEnum controls when autocomplete
is not turned off via the config or ui schema options.

We reuse the autocomplete component of '@material-ui/lab'. The
new renderers are kept separate in 'extendedMaterialRenderers' as
we don't want to force this dependency to all consumers.