Introduce validation mode
Adds validationMode to the core state supporting
 * Validation & showing the validation result (default)
 * Validation & hiding the validation result
 * No validation

'Validation & showing the validation result' is the default mode and
represents the previous mode. 'Validation & hiding the validation
result' will still keep the errors in the core state up to date, however
they are not given to the respective renderers by default. 'No
validation' will short-circuit the validation, resulting in an always
empty errors array in the core state.

The mode can be set as part of the init action and can later be updated
with the separate setValidationMode action. The action is exposed as a
separate prop in React and as a separate method in the Angular service.

The new modes can be used for various use cases, e.g.
 * save performance by never validating on client side
 * show validation errors separately from the form
 * only show validation errors on submit or after first edit