Use existing Ajv instance in all places
Previously a separate Ajv instance was used to evaluate schemas for the
rule support. We now reuse the instance from the state / context for this
evaluation. With this change the same Ajv instance is used for all use cases.

This change enables full Ajv customization support as there are no longer
"hardcoded" instances in use.