Refactor "Clear input" button
The 'Clear input' icon overlayed the input text in
such a way that part of the text wasn't visible, even
when the icon itself wasn't shown.

We now use the 'display' css property instead of
'visibility' to avoid this issue.

Additionally we set a background color to the center
of the button so that the button is clearly visible
against the underlying text. As we had to choose
*some* color we made it configurable with the custom
theme variable 'jsonforms.input.delete.background'.