Move onChange handling to StateProvider
Move onChange handling from JsonFormsDispatch to StateProvider as it
also manages the source of truth, i.e. the JsonForms core containing
the emitted data and errors.

Also use `useEffect` instead of `useLayoutEffect` as the latter one is
not needed here. This also removes warnings when using JsonForms in SSR
settings.