Add manual trigger of changeDetector in Master Renderer
When the data array of the list master detail renderer is empty,
then changing the data does not rerender the ui.
To fix this the changeDetector is manually triggered.
This error can only be reproduced in the seed,
not from the dev environment.
Furthermore the readonly handling was fixed.

Fix #1667