Integrate (ui) schema based control enablement
The JSON Forms core utilities will now also consider the JSON Schema
'readOnly' property as well as 'readonly' and 'readOnly' properties
defined in ui schemas.

The algorithm checks in the following order whether the form wide
readonly flag is set, an enablement rule is defined for the ui schema
element, ui schema or config options are set and whether the JSON Schema
specifies 'readOnly'. If none of them apply the enablement of the parent
is used.

Includes example and test cases.