Only emit unique events in React
Adds an event cache to the React integration for emitted events.
Skips emitting events which were already emitted very recently.

This fixes an endless rendering loop issue in cases where the user
dispatched many change events in a small amount of time, e.g. via
Browser autofill features. When that happened and the user also
fed back the emitted data to JSON Forms, the user could end up in
an endless update cycle of type A -> B -> A -> B... . This is now
avoided.