Escape special chars in JSON Pointers
Escape special characters in JSON Pointers to support forward slashes in
attribute names.

JSON Schema contains JSON Pointers in "$ref" attributes and UI Schemas in the
"scope" attributes. As the forward slash is used a separator in in JSON
Pointers, but attributes can contain forward slashes in their name, these
pointers need to be encoded and decoded.

This commit adds an "encode" utility to escape JSON Pointers. Once we convert
to a data path or want to access the raw name, "decode" can be used.

Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>