Offer breakpoint option for horizontal layouts
The horizontal layout renderer now looks for the new 'breakHorizontal' UI Schema option.
This option can either be 'xs', 'sm', 'md', 'lg' or 'xl' which corresponds to the
breakpoint support of Vuetify. If no breaking behavior is wanted the option can be
anything else, e.g. 'false'.

The layout will render vertically if the given breakpoint, or a smaller one, is hit.
Thefore the layout will always render vertically when `xl` is given, while option 'xs'
will only break the layout on the smallest view ports.