Add support for 'then/else' paths
Schema paths containing 'then' or 'else' were not transformed correctly to their data
paths. This is now fixed.

The JSON Forms resolver already allowed to omit 'oneOf', 'allOf' and 'anyOf' within UI
Schema scopes. The resolver will now do the same for 'then' and 'else' segments.

For this the resolving algorithm was rewritten to a depth first search. First trying to
resolve the paths as they were handed over and, if not successful, checking for omitted
segments.

Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>