Add external error integration
Additional errors can now be mixed in with the errors managed by JSON Forms itself. This
mechanism can for example be used to integrate backend errors into forms.

The additional errors are handed over via the respective root components. They are merged
with the internal errors, however they are not affected by the current validation mode.

Implements #1926

Contributed on behalf of STMicroelectronics

Signed-off-by: Alexandra Buzila <abuzila@eclipsesource.com>