Improve vanilla ArrayControl
The vanilla ArrayControl was wrapping the `addItem` in a function,
whereas the `addItem` function is already wrapped. So the `onClick`
can directly be bound to `addItem`. The wrong binding lead to the
add functionality not working.

Increase rank of the ArrayControl so that it behaves the same as
for the react material renderers.

Add example to demonstrate different selections of renderers based
on the existince of a const property in the schema.

Fix wrong selection of array layout if type is missing.

Fixes #1948