Improve React Material date/time input handling
Whenever the user entered an invalid date/time string, the renderers stored `undefined`
in the data, i.e. the attribute was deleted. This is now changed to pass through the
invalid string which is a much more flexible behavior.

However dayjs is actually very lenient when parsing strings and tries its best to parse
*something* out of any given string. This can have the effect that what the user entered
is not necessarily what is actually stored in the data. To improve UX the date/time
inputs now always reflect the actual stored data once they lose focus.

Co-authored-by: Stefan Dirix <sdirix@eclipsesource.com>