Translate UI Schema elements
'Group', 'Category' and 'Label' UI Schema elements are now translatable via bindings.

'mapStateToLayoutProps' is enhanced to check for 'label' or 'i18n' attributes within the
UI Schema element. If that's the case the translator will be invoked and the label handed
over which is then used by the group renderers.

The categorization renderers access the new utility 'deriveLabelForUISchemaElement' to
translate the labels of their categories. This is done within the respective renderer
code.

For labels a new `mapStateToLabelProps` mapper is added which provides access to a 'text'
property. Corresponding bindings are added to all renderer sets and used by the
respective renderers.

Also improves some typings in core and streamlines prop handling in the React bindings.