Provide context object to Testers
Changes the Tester 'rootSchema' parameter to the new 'TesterContext' object. It contains
the 'rootSchema' and the form wide config object. The 'TesterContext' might be further
enhanced in the future.

Also cleans up the 'mapStateToJsonFormsRendererProps' by removing the UI Schema
generation and properly integrating it within the React bindings.

Additionally upgrades the '@vue/test-utils' and integrates the config object into the
dispatch renderer.

Fixes #1970