Update to JSON Forms 3.0.0-rc.0 and Vue 2.7
Updates to JSON Forms 3.0.0-rc.0 which is based on Vue 2.7 which now contains the
Composition API by default. Adapts the renderers and example application where necessary.

Also requires some casts to 'any' as we run into problems with the Vue 2.7 types in some
cases (excessive stack depth when comparing types).