Resolve schemas in combinator mappings
Within the combinator mappings we try to determine the best fitting schema so that the
renderer can decide which tab to show. With the removal of the auto-resolving this no
longer worked in cases where the combinator contained references.

The situation is now improved by resolving references on the combinator level. This
should help in many use cases but still does not restore the old behavior completely.
Fully checking and resolving the whole sub schema is out of the scope with the new
simplified resolving architecture. If that is needed the user either needs use a custom
renderer or restore the old behavior by fully resolving their JSON Schema before handing
it over to JSON Forms, for example via json-refs.

This commit also removes the full resolving of the JSON Schema within the example
application which did mask this bug.