Improve angular performance
Improve performance by avoiding unnecessary cloning of the form wide state and avoiding
unnecessary recalculation of the child render props, therefore avoiding unnecessary
change detection cycles in children.

When querying the state from the `JsonFormsAngularService` via its `getState` method, the
whole state is cloned. This lead to a heavy performance penalty because the
`JsonFormsAbstractControl` queried the state on every change detection cycle to get the
global config. The `NumberControlRenderer` queried the state on every change detection
cycle to get the locale. Therefore new methods to receive the config and locale are added
which avoid cloning the whole state.

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>