Add examples app with bundled react vanilla, material examples
This introduces a new private package `examples-app` to create a simple web page
that ultimately has child pages for every example.

In this initial commit, only the react vanilla and material examples are added.
Therefore, they are bundled with rollup to be able to run as standalone pages.

- Add rollup configs to bundle react vanilla and material examples
- Add required dev deps and upgrade rollup dev dep to ^2.78.0 for all packages
- change material deps to ~5.2.2 because latest 5.x versions break api for Autocomplete
- Add private package `examples-app` with a script to aggregate the examples
- Add root level npm script to create the examples app
- regenerate package-lock.json

Part of #1706