Consider oneOf enum in array testers
Whenever we encounter an array we run a heuristic which determines whether the
'items' are shallow objects/values or some more complex objects. This is typically used
to determine whether a compact table renderer or more complex element renderer
shall be used.

This heuristic is extended to consider 'oneOf's which model an enum to be suitable for
table renderers.