Do not filter out errors for oneOf enums
The errorsAt function, used to filter the errors that belong to a
control, ignores the errors AJV produces for oneOf schemas. Most of
these errors should not be shown in the UI (e.g. when the oneOf property
does not match exactly one subschema), however the errors produced
directly for oneOf enums are an exception.

Validation errors for oneOf enum controls are valid and should be shown
in the UI.

Contributed on behalf of STMicroelectronics