fix: properly match AJV paths with slashes
AJV error paths are JSON Pointers. These paths are converted to dot-notation in JSON
Forms. However escaped JSON Pointer characters were not decoded. Therefore errors for
properties containing one of the escaped characters were not properly matched.

Background information: In JSON Pointers two characters need to be escaped: '~' is
converted to '~0' and '/' is converted to '~1'. Therefore properties containing either
'~' or '/' were not properly matched to errors before this fix.

Fixes #2146