build: upgrade stack to pnpm, Node 16 and Lerna 6
Use pnpm 8.5+ as package manager to be used on Node 16.13+. Also upgrades Lerna to v6
with nx backend.

As part of the repository overhaul the Vue 2 packages were removed. Starting with JSON
Forms 3.1 they are no longer maintained.

All packages now properly declare all their dependencies. A number of dependencies were
updated in the process, e.g. jest and typedoc.

Other changes:
 - Increases base lib version from ES6 to ES2016.
 - Update Readme, scripts and Github workflows
 - Increase RAM usage for test executions
 
Note: The version constraint of pnpm in the root package.json is set to ">=7.13.4" as
otherwise the Netlify build fails immediatly. It should be set to ">=8.6.2".

Closes #2110