Build Angular packages with ng-packagr (#2236)
This replaces the use of ngc with ng-packagr when building the angular packages.

With this, the libraries don't provide a cjs output anymore, only (partially compiled) *esm.

In order to tidy the output, the sources have been moved to a library sub folder and exposed via a commonly used public_api file.

The example for the angular material renderer has also been modified to import the output of the library instead of the sources. This allows to actually validate that the package can be loaded and fully compiled (rather than rebuilding everything, the lib + the app, all at once). Therefore, the use of rollup has been dropped in favor of the internal angular browser builder.

This also adds `@angular/core` as dev dependency to root package.json to facilitate the new build
working with angular-material's `dev` npm script from within the package.

---------

Co-authored-by: Lucas Koehler <lkoehler@eclipsesource.com>