vue-vuetify: Extract control entries and fix usage instructions
fix #2378

The registry entries for the renderers were configured in the same files
as the renderers. This lead to the entries being removed due to tree shaking
during production builds using esbuild (as used by Vite).

This extracts all entries to separate files and improves the README to no
longer import the renderers asynchronously .