Make HTML ID generation fast
Replaces HTML ID generation with a simple monotonic sequence to avoid
several expensive hash lookups for every generated ID.

This entirely eliminates the chance for collisions under regular
circumstances, and the default seed even performs better in some
adverse circumstances, like multiple instances of the same library
sharing an ID namespace on the same page.

The only potential downside is that IDs might not be particularly
semantically helpful, but that can be mitigated by just appending that
information instead.

This might be a breaking change if this module is considered a public
API.