angular: Migrate to standalone components
Starting with Angular 19, standalone components are the new default and
non-standalone components are discouraged producing a lint error.

This migrates all components of the Angular and Angular Material packages
to be standalone. Backwards compatibility is maintained by keeping existing
NG modules which now import and export all components.
Thus, users can use the modules as before.